344691 Work around MD5 problem for signed JARs
The fix in this patch consists of adding the flag
IPublisherInfo.A_NO_MD5 while generating the p2 metadata.

To support the fix, this patch also contains a new mojo that
allows to verify md5 sums and signatures in built repositories.
This VerifyIntegrityRepositoryMojo is not integrated into the
default build lifecycle but used to add md5 verification to the
jarsigning integration test. (Without the fix above, the wrong
md5 sum would show up in the integration test result).

Bug: 344691 Invalid MD5 sums for signed jars in eclipse-repository