364486 better handle remote p2 repository access problems
When remote p2 repository is not accessible and there is no local cache,
fail the build fast and with a clear indication of the root cause.

When remote p2 repository is not accessible but there is local cache,
warn about the problem with the remote repository but use the cache.

Signed-off-by: Igor Fedorenko <igor@ifedorenko.com>