365012 protect p2 runtime extraction with lock
To achieve this, we have to move the
FileLockServiceImpl from org.eclipse.tycho.p2.resolver.impl
(running in embedded equinox) to tycho-core.
Then we enable registering OSGi services
in EquinoxServiceFactory and register the
file locking service on OSGi framework startup.
This way the same object can be used as plexus
component as well as OSGi service.