[refactor] limit cache size and make it an LRU cache
Instead of HashMap, use a least recently used (LRU)
cache based on LinkedHashMap for keeping
MANIFEST and build.properties in memory.
Limit size to 50 objects which is enough for
keeping all objects in-memory for most reactors but
will prevent excessive memory use for very big
reactors. Performance should still be OK for bigger
reactors since we keep the
50 most recently used objects in memory.