367701 documented and fixed IT project, hooked into java test suite
Added dependency on org.eclipse.core.runtime to 'testbundle' project,
which is required to compile the test class.

Removed x-installation:=greedy optional dependency attribute, which
effectively obscured the problem with indirect dependencies of
reactor projects.

Added the project to the test suite, which now fails due to the
presence of indirect optional dependency.

Signed-off-by: Igor Fedorenko <igor@ifedorenko.com>