375160 Add name and vendor attributes to source bundle generation.
When a source bundle is generated, the bundle's name and vendor attributes
should be passed from the original bundle. The bundle's name should have
"Source" appended to the end to distinguish it from the original bundle.
Additionally, the Bundle-Localization attribute must be checked in case
certain attributes are stored in a separate configuration file.

Change-Id: I7982d0578ca9d5ebe61bea6adcd78d35683e0d6f