375160 Add name and vendor attributes to source bundle generation.
When a source bundle is generated, the bundle's name and vendor attributes
should be passed from the original bundle. For bundles implementing
Bundle-Localization, the property file to be used for the source bundles
must be located in the same directory as the original property file, and
have "-src" appended to the name.

Change-Id: I7982d0578ca9d5ebe61bea6adcd78d35683e0d6f