Update patch set 1
Patch Set 1:

I am not convinced we need all this complexity. <disableMirrors> is about build environment configuration and as such does not belong to pom.xml. A conventional property in settings.xml is a better solution. 

I am not going to veto this change, but I think we should remove <disableMirrors>

Patch-set: 1