Update patch set 2
Patch Set 2:

Maybe it's my lack of experience with Gerrit, but each bullet from the comment message looks like a good candidate for a separate change (after all, you wanted to highlight them as separate items).

Again, this is not to veto this change, just a general observation. I am still getting used to Gerrit, so apologies if my comments appear random and inappropriate. ;-)

Patch-set: 2