Update patch set 2
Patch Set 2: Abandoned

I was just going to write how it was not reasonably possible to split up this change in small increments, when I realized that it is. The changes from this commit are now contained in change https://git.eclipse.org/r/#/c/5604/ and its predecessors.

Although I agree that it is good to have small-ish yet self-contained commits in the commit history, it was quite a bit of effort to produce them. IMHO we need to reconsider what to do if we get a similar situation in the future.

Patch-set: 2