Update patch set 1
Patch Set 1: Do not submit

(3 inline comments)

Great patch, but unfortunately it is not quite correct in some corner cases. This is why we can't commit it in the current state. Still, I am convinced that a cache is the right way to go, so I'll be working on it to fix the corner cases and to bring it in.

Patch-set: 1