Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

(4 inline comments)

I think that the changes needed in p2 are doable in p2, so I would prefer that you propose a patch there first before creating copies in Tycho.

But as the vote says, this is just a non-binding, personal preference; you are free to overrule this vote by voting +2.

Patch-set: 1