375160 Add name and vendor attributes to source bundle manifest
When a source bundle is generated, the bundle's name and vendor
attributes should be passed from the original bundle.
Bundle-Name and Bundle-Vendor values of the original bundle are
i18n-resolved and resolved values (with " Source" appended for
Bundle-Name) are used to generate OSGI-INF/l10n/bundle-src.properties
for the source bundle.

To allow overriding generation, if OSGI-INF/l10n/bundle-src.properties
is included in src.includes, this file will not be generated.

Default value for Bundle-Name (if missing) is Bundle-SymbolicName.
Default value for Bundle-Vendor (if missing) is "unknown".
Warnings are logged in both cases.

Change-Id: I7982d0578ca9d5ebe61bea6adcd78d35683e0d6f