Update patch set 2
Patch Set 2: Verified; IP review completed

Honestly, I have trouble understanding the contract of P2ResolutionResult. Plus I'm not convinced if testing against this interface is a good idea, given that it is not used by the Tycho code on the Maven (where a DefaultTargetPlatform/DependencyArtifacts instance is used)

So I just tested on the level that is clear: there should not be the "xyz is not part of the target platform" error.

Patch-set: 2