Update patch set 1
Patch Set 1:

Resolver configuration and input codata is spread over too many places. Some of it is accumulated in P2ResolverImpl, some in ResolutionContext, some in P2TargetPlatform (how many instances of it we have?) and some is passed as parameters of P2ResolverImpl#resolveProject. I think we should consolidate this before it gets out of hand. 

For example, we can move all resolution related state from P2ResolverImpl to ResolutionContext and pass ResolutionContext as P2ResolverImpl #resolveProject and similar methods.

Otherwise the change looks reasonable, so feel free to merge it as is and we'll refactor P2ResolverImpl/ResolutionContext later.

Patch-set: 1