Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

This change absolutely makes sense, can can be submitted IMHO.

Still, I'll need to show you some of the work I have done to clean up the publishing code (e.g. in order to get rid of the special handling of root files). This is only POC quality, but I'll upload a change to let you jugde if this could be useful for you.

Patch-set: 1