m2e-tycho workspace dependency resolution support
When tycho build is invoked from within m2e-tycho workspace
and workspace dependency resolution is enabled for the launch
configration, tycho p2 runtime and external bundles will be
resolved according to workspace configuration.

This allows execution of Tycho build directly from Tycho
development workspace, without the need to first run mvn
intall or deploy.

Implementation relies on corresponding support introduced in
m2e-tycho 0.7 and is dormant when required m2e-tycho version
is not installed or when workspace dependency resolution is
not enabled for the launch configuration.

Change-Id: I8dc9cea66bcdbb3712288e562b877b35b26c585c
Signed-off-by: Igor Fedorenko <igor@ifedorenko.com>