Update patch set 1
Patch Set 1: Verified; Looks good to me, approved; IP review completed

I found a couple of problems with IPropertyAdvice:
a) They currently only have an effect on SimpleArtifactDescriptors, a type that we don't want to use. 

b) In order to fix this, there is the idea to change the signature of IPropertyAdvice to 

IArtifactDescriptor setArtifactProperties(IInstallableUnit,
IArtifactDescriptor)

(see https://bugs.eclipse.org/bugs/show_bug.cgi?id=294104 ), i.e. to have the advice directly edit the descriptor. (Interestingly, this is exactly what our MavenPropertiesAdvice does). So I also could do this, but without the signature change, this feels kind of hacky. (And I don't feel like fixing that p2 bug right now...)

c) Setting properties in IArtifactRepository.createArtifactDescriptor is recommended in a different p2 bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=254652#c3
IMHO the biggest advantage of this (i.e. the current) approach is that it is possible to write to an artifact repository without being forced to use the PublisherHelper. I always found this really weak of the IArtifactRepository APIs that they are close to impossible to use witout a large set of additional tools. At least we are not ading to this...

Patch-set: 1