355367 property to control if local metadata is considered
Add property "tycho.considerLocal" which allows
to control whether p2 metadata of artifacts installed
in the local maven repository by previous tycho builds
should be considered by the resolver or not:

- if not set, behaviour is unchanged (local metadata
  is considered by the resolver, but warn if local-only IUs
  were used)
- if set to "true", consider local metadata but do not warn
  about usage of locally installed IUs
- if set to false, ignore local-only metadata

The property can be specified on the commandline
or in settings.xml.
While it is technically possible to set it in pom.xml also,
this is not recommended and not officially supported.

Using -Dtycho.considerLocal=false can be desirable
if you want to isolate your build from bundles installed
in the local maven repo by previous tycho builds
on the same machine.

Change-Id: Id938d4b9484302ffaa000947b6d8adb43f0a2469