Update patch set 2
Patch Set 2: (1 inline comment)

I don't know if this was intentional or not, but I believe it is a mistake to allow projects to decide if locally installed artifacts should or should not be considered by the build system. Besides, this change allows different modules of the same reactor to have different configuration, either on purpose or by mistake. Good luck troubleshooting such builds. 

In any case, I can show how to get properties from settings.xlm only, but feel free to merge this change as is if you actually want to be able to configure this in pom.xml.

Patch-set: 2