Update patch set 2
Patch Set 2: (1 inline comment)

> I don't know if this was intentional or not, but I believe it is a mistake to allow projects to decide if locally installed artifacts should or should not be considered by the build system.

I agree that this use case is questionable, but I think the main guidance should be how other Maven plug-ins handle system properties that have an effect by convention. I would dislike if we invented the "system property that can only be set via settings.xml and command line but not via <properties>" in Maven.

So if a project really thinks it needs to set the conventional property in its POM, the builder can always still override it on the command line, right? IMHO this is the lesser of two evils.

> Besides, this change allows different modules of the same reactor to have different configuration, either on purpose or by mistake. Good luck troubleshooting such builds.

I don't see a big issue here: the target platform may be different of every module today (so there is no fundamentally new complexity), and obviously there needs to be debug output during target platform computation about the effective value of the conventional property.

Patch-set: 2