Update patch set 2
Patch Set 2:

I agree that projects should not set this property.

But is it critical if they would do it? If you think it is, then we need to prevent this technically and hence need to introduce the (previously not existing concept in Maven) of "reactor-global properties".

I can live with that because we don't need to document that the new property behaves differently than the properties that people know.

Patch-set: 2