376973 introduced tycho.disableP2Mirrors session property
tycho.disableP2Mirrors can be specified either on command line
as -D parameter or in settings.xml. Use of p2 mirrors is enabled
by default.

The old target-platform-configuration <disableP2Mirrors> is not
supported any more and the build fails fast and with clear error
message when this configuration element is present in pom.xml.

Change-Id: Ibd5fb1f044293f759752fb0f59dc64a7932f9282
Signed-off-by: Igor Fedorenko <igor@ifedorenko.com>