Update patch set 4
Patch Set 4:

> Remove requirements on jre from IUs generated by Tycho. This will properly refrect the lack of EE support in existing p2 clients without the need to pollute Tycho-generated repositories with garbage IUs. 

What do you mean by garbage IUs? The "regular" a.jre.xxx IUs generated by the JREAction, or only the empty a.jre.javase/1.6.0 IUs that Tycho creates as workaround for bug 370502? The latter will never be included in Tycho-generated p2 repositories.

And for reference: opened bug 387796 for breaking the hard dependency from products to an 'a.jre' IUs.

Patch-set: 4