388629 removed eclipse.inf from pack200 mojos
Due to p2 bug 387557, it is not possible to install bundles signed
and packed by Tycho. As a workaround, do not generate eclipse.inf
file when packing bundles with tycho and fail the build fast for
bundles that already include eclipse.inf.

Note that Tycho provided support for eclipse.inf for backwards
compatibility with existing projects, Tycho itself does not need
the file to properly sign&pack bundles.

Change-Id: Ieeb28a512dbc64694593dc50a26ecdb7ffaa8984
Signed-off-by: Igor Fedorenko <igor@ifedorenko.com>