Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

I have to admit that "tycho.considerLocal" sounds way snappier than "tycho.includeLocal". Maybe we just leave it as it is. 

Just in case we leave everything as it is, I pulled out the refactorings into as separate commit.

Patch-set: 2