Update patch set 7
Patch Set 7:

yes this is on my list for 0.16. Just came back from vacation so I will need to catch up.

regarding the phase for validation, currently it makes no sense to postpone it e.g. to verify. This is because tycho reads the target file and resolves the target platform in the early pre-lifecycle (afterProjectsRead()) already.

Patch-set: 7