380872 sourceref provider for jgit
Implement SourceReferenceProvider
for SCM type "git".
Use jgit to find the git repo root
and calculate path of project basedir
relative to the repo root.
Add it as ";path=..." header attribute.
Resolve HEAD commit and add it
as ";commitId=<SHA1>" attribute
If HEAD is also a tag, add it as
";tag=..." attribute.

To enable Eclipse-SourceReferences generation
for projects using git, you have to:

1. declare ${tycho.scmUrl} in your (parent)
   pom using maven SCM format [1]

2. Configure tycho-packaging-plugin to
   generate source reference headers and
   add a dependency to tycho-sourceref-jgit:

<plugin>
 <groupId>org.eclipse.tycho</groupId>
 <artifactId>tycho-packaging-plugin</artifactId>
 <version>${tycho-version}</version>
 <configuration>
  <sourceReferences>
   <generate>true</generate>
  </sourceReferences>
 </configuration>
 <dependencies>
  <dependency>
   <groupId>org.eclipse.tycho.extras</groupId>
   <artifactId>tycho-sourceref-jgit</artifactId>
   <version>${tycho-extras-version}</version>
  </dependency>
 </dependencies>
</plugin>

[1] http://maven.apache.org/scm/scm-url-format.html

Change-Id: I447d7dc244e6d73c797bf43877b13f9e49fd0ded