392193 Omit check for duplicate IUs in TestMojo
When computing the target platform in TestMojo, omit the check for
duplicate IUs produced in the reactor. Duplicate a.jre.* IUs (which are
in fact identical) can be produced if several products are published.
This is a workaround for 0.16.0, a better solution for 0.17.0 should be
addressed with fix for bug 392320.

Change-Id: Ia4282b6682126cedb92cb161e889048db493de4d
Bug: 392193