Update patch set 2
Patch Set 2: Looks good to me, approved; IP review completed

Not nice in how many layers we need to add the boolean switch, but I don't see a better solution either.

Added a comment to the integration test. Test may be merged with other product integration tests, but we don't need to do this now.

Patch-set: 2