Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

We can't validate-id and validate-version of the products as long as they are in a eclipse-repository. We'd need the eclipse-product packaging type for this.

In any case, the validation topic is unrelated and should not be mixed with the qualifier-aggregation change.

Patch-set: 1