Update patch set 1
Patch Set 1:

As mentioned in the bug report, this patch was not meant to be production ready.
As it was originally requested I put it here so we could discuss the scope of the changes implied by the addition of the fork mode, and thus know whether or not, given the current status of surefire, there was a chance for this feature (implemented correctly on top of this patch) to be included in the short term, or if we had wait until a later time.

Patch-set: 1