395281 resolve runtime from explicitly configured p2 repos only
As a build plugin, the runtime for eclipserun-plugin can't be built in the same reactor in which it is used.

Instead of reusing the target platform of the currently built  project, use independently configured p2 repositories to resolve
the eclipse runtime to be launched.

also, add a skip parameter to the mojo and fix return code handling so the build fails if the forked process failed.

Default dependencies to launcher jar etc. can now be switched off.

Bug: 395281
Change-Id: I69012098f5f8142d4be9c5a14118266706572479