Update patch set 1
Patch Set 1:

I think it would be even better to delete it straight away. Just removing it from the default lifecycle will make Maven (silently) ignore any configuration that people have made for this goal. 

I think that explicit calls to the goal (which is the reason for keeping it) will be less common that the case above.

Patch-set: 1