Update patch set 7
Patch Set 7:

This change is a bit large. I'll split out the parts that introduce the new (p2) API for streaming artifacts (around IAartifactSink) into a separate change.

Most TODOs that need to be fixed before this can be submitted are addressed. Exceptions: pack200 mirroring, logging, and the TargetPlatformBundlePublisher artifact repo impl.

Patch-set: 7