410418 add binary feature dependency to dependency only metadata
Source features include the corresponding binary features
by default. This was not taken into acount when generating
"dependency-only" metadata for source features and could lead
to unresolved dependency errors when packing the feature jar.
Note: This is a stop-gap solution until we fix bug 353889.
AbstractArtifactDependencyWalker does still not honor source
feature classifiers when traversing dependencies, but since
it is only used to compute version qualifiers
(and these are the same for source and binary feature)
I am proposing this workaround.

Bug: 410418
Change-Id: I86e4d74ae03a5537eb8b070399cd745e6e5a76c8
Signed-off-by: Jan Sievers <jan.sievers@sap.com>