Update patch set 3
Patch Set 3:

Ok, I think I got it. I did not understood it well at first read.

It seems that the property I'd like to set from command-line "work" which is the place where the application is expected to be installed. Is that right?
If so, can you please make this "work" and "provisionInstallation" properties configurable via command-line so it would avoid creation of profiles for occasional tests?

Patch-set: 3