Update patch set 4
Patch Set 4:

I guess the launcher has to available in the target platform of the test module. The tycho demo example itp04-rcp and eclipse classic SDK both work for me when running tests on a pre-provisioned installation with kepler p2 repo in the pom. Can't tell what's different in JBDS. will need an example to reproduce.

Patch-set: 4