Update patch set 4
Patch Set 4:

@Tobias, we're speaking here of a use-case that is currently not implemented, that's why it's rarely needed! It can be tycho.surefire.profileName if you want.

The profileName property will need to be set almost everytime the provisionedInstallation will be set to true. For example, test against an EPP package need to override this property to "epp.package.rcp" for example.

I would really like to avoid polluting my pom with 7 lines of XML just for setting a property in a Mojo, and so far, the main use-case is to have quick configuration of which application hosts the test via command-line.

Patch-set: 4