Update patch set 4
Patch Set 4:

I'm just not convinced that having a global property has a cost for the community.
I agree we can find a better name which reduces risk of collision with other mojos.
I've seen most Mojos having to add some properties later to ease configuration. Adding properties seems to be a common pattern while a Mojo is getting maturity. The best recent example I have in mind is the cbi-signing-plugin which required 2 profiles and many lines of configuration to use it in and out of Eclipse.org, and all those profiles could be removed just by introduction of a cbi.signing.skip property that can be set from command-line.

A property called "installation" or "profileName" can have a risk, but tycho.surefire.installation and tycho surefire.profileName have no cost/risk.

Patch-set: 4