Update patch set 4
Patch Set 4:

"Adding properties seems to be a common pattern while a Mojo is getting maturity"

Exactly. This functionality is brand new and still needs to be proven. We can add the properties later if it turns out that many users need to change it from the command line. If you want, create an enhancement request and get people to vote for it.

Patch-set: 4