Update patch set 4
Patch Set 4:

What's the benefit of waiting before making a field configurable via command-line?

The whole use-case which led to writing the first patch is based on the assumption that this needs to be configurable via command-line so it's easy to give it a try on any existing module without requiring any change in pom files.

Anyway, let's start by no property if you're totally opposed to that, and as you said, I'll request them next later (next week ;) )

PS: I'm not against criticism on the feature or the implementation, I just don't understand the concerns caused setting an expression on a field.

Patch-set: 4