Update patch set 4
Patch Set 4:

I suggest the following commit to be added to this patchset: https://github.com/mickaelistria/tycho/commit/90111786a4b3199e1cbb74702509c90facf2228a

It introduces an intermediary bundle for provisioned application which is defining dependency on all required bundles with an open version range, and install this bundle instead of installing all bundles explicitly.
The benefit is that it will rely on p2 to make an optimal resolution, so it allows to run tests on a target installation which does not have the same platform version (org.eclipse.osgi, org.eclipse.core.runtime, org.eclipse.equinox.laucncher, org.eclipse.u) as the one configured for the module build.

Patch-set: 4