Update patch set 4
Patch Set 4:

"It introduces an intermediary bundle for provisioned application which is defining dependency on all required bundles with an open version range, and install this bundle instead of installing all bundles explicitly"

This is probably a good idea, but is not needed for this change to become ready to be committed. So let's keep this proposal in a separate commit. Divide and conquer ;-)

Patch-set: 4
Reviewer: Gerrit User 489 <489@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0