Update patch set 2
Patch Set 2:

(3 comments)

Thanks for the review.
If you have time to give answers about the suggestions and questions I put in my comments, I'll work on a new version of this patch with those small improvements, so it will be less work for you on Friday.

Patch-set: 2
Label: Verified=0