Update patch set 3
Patch Set 3:

(1 comment)

I changed things around so that I'm no longer overwriting the original update site (not sure why I didn't think to do this the last time to be honest).

Instead if site.xml is found it'll do what it used to do and throw a warning.

If category.xml is found it'll create an eclipse-repository pom instead.

Patch-set: 3
Label: Verified=0