Update patch set 2
Patch Set 2: Verified+1

@Igor: Yes, all tycho-extras integration tests pass. (I have a local Jenkins set up for running all integration tests on proposed changes.)

According to the commit messages which introduced the IArtifactFacade usage in the target platform, this was for needed for getting the final metadata into the TP. Final metadata still works, I just changed that the metadata is stored per project and not per project-sub-artifact. (Note that the sub-artifact information can anyway be recovered from the "maven-classifier" attribute in the IU.)

If there is in fact something not working any more, could you contribute an integration test which detects the problem? Then, I'll see how I can fix this.

Patch-set: 2
Label: Verified=+1