Update patch set 1
Patch Set 1:

I don't have too strong an opinion about this change, but I don't think it improves the code -- (very few) more lines of code and the logic is (slightly) harder to follow I think. But feel free to merge, the change is tiny and to the private classes, so the scope of the change is trivial.

Patch-set: 1
Reviewer: Gerrit User 583 <583@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0