Update patch set 1
Patch Set 1:

The logic is hard to understand, now and before. OOP's answer to this complexity is separation of concerns and localizing things. This is better than before. But this comes at a price of a new interface (the HierarchyFormatter's methods&members). 

The new code is more verbose because the strings which are inserted have names now. I think that this makes the new code better.

But really, it is still hard to read. BTW, the tests are a lot easier to read, and as long as they pass, one doesn't absolutely have to understand the production code ;-)

Patch-set: 1