427556 use commons-exec instead of plexus-utils
The process watchdog will try to destroy the process
right away if the timeout is exceeded instead of
during Maven VM shutdown only.
Also, commons-exec will not use an intermediate shell to
fork the process (this was only used for historic reasons
in pre-JDK 1.3(!) times, when there was no java.lang.Process API to set
the
working dir of the forked process).
Remove argument escaping code as this was only needed due to 
the intermediate shell process which could mangle arguments
e.g. with spaces. All arguments are passed as-is to 
Process.execute(String[] args), no explicit escaping necessary.

Note that this still does not guarantee that
Process.destroy() will work in all cases.

Filed CQ 7907 for Apache Commons Exec.

Bug: 427556
Change-Id: Ia151ca28892b9f6f3f351bd134834904ba7baeec
Signed-off-by: Jan Sievers <jan.sievers@sap.com>