427556 use commons-exec instead of plexus-utils
plexus-utils uses an intermediate shell
process which effectively prevents
process destruction of the forked JVM.
For details see https://jira.codehaus.org/browse/PLXUTILS-40

Use commons-exec instead which does not use an intermediate
shell and provides
a process watchdog. This will try to destroy the process
if timeout is exceeded.
In addition to this, also register a process destroyer as VM shutdown
hook to cover the case where the build process is killed
(best effort to avoid orphaned child VMs).

Remove argument escaping code as this was only needed due to
the intermediate shell process which could mangle arguments
e.g. with spaces. All arguments (except TestMojo#argLine and
TestMojo#appArgLine) are passed
as-is to
Process.execute(String[] args), no explicit quoting or
escaping necessary.

Note that this still does not guarantee that
java.lang.Process.destroy() will work in all cases and on all
platforms.
Also, if the build process is killed forcefully with -9
(KILL signal instead of default TERM signal),
JVM shutdown hooks will not be executed and hence the child
process will not be destroyed.

Filed CQ 7907 for Apache Commons Exec.

Bug: 427556
Change-Id: Ia151ca28892b9f6f3f351bd134834904ba7baeec
Signed-off-by: Jan Sievers <jan.sievers@sap.com>