[DRAFT] 427556 use commons-exec instead of plexus-utils
The process watchdog will destroy the process
right away if the timeout is exceeded instead of
during Maven VM shutdown only.


Note that this still does not guarantee that
Process.destroy() will work in all cases.

TODO if we want to go this way we need to file
a CQ for commons-exec.
Also, there are probably other places in tycho where we fork processes
with a timeout, not sure if those should be changed too.


Bug: 427556
Change-Id: Ia151ca28892b9f6f3f351ad1b4834804ba7baeec
Signed-off-by: Jan Sievers <jan.sievers@sap.com>