[cleanup]Use Arrays As List for Performance
The class java.util.Arrays has a asList method that should be use when
you want to create a new List from an array of objects. It is faster
than executing a loop to copy all the elements of the array one by one.
Also remove useless "Suppress Warning" annotations

Change-Id: I19eabd52ffc773680b70b0fdfa4cbe72254c3c71
Signed-off-by: Aurelien Pupier <aurelien.pupier@bonitasoft.com>