Update patch set 1
Patch Set 1:

(1 comment)

AFAIK, Tycho never serializes exceptions, so this field is not necessary.

Right now, I don't think that we have a concept of how we want to handle this. So I'd rather not just (randomly?) change the UUID in one class, but rather come up with a concept and then follow that concept consistently throughout the Tycho code.

However this discussion is rather something for tycho-users or a bug.

Patch-set: 1
Reviewer: Gerrit User 489 <489@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0