Update patch set 1
Patch Set 1:

Using the wiring API would of course be a lot better. If the migration is not hard to do, we should definitely do this. I've opened a new bug report for this: http://bugs.eclip.se/432137

AFAIK, the wiring API already exists in Equinox from Kepler, so the wiring API migration could even be done before the Luna migration.

Patch-set: 1
Label: Verified=0