Update patch set 1
Patch Set 1:

we have > 300 of these 

https://dev.eclipse.org/sonar/drilldown/issues/84397?&rule=squid%3AS00121&rule_sev=MAJOR&severity=MAJOR

I don't consider this a major problem.

However if we want to change it, either we fix them all (and change save actions in all projects to automatically add braces so we never see this warning again) or we leave it as is.

Patch-set: 1
Reviewer: Gerrit User 434 <434@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0