429809 symlink target computed always relative to archive root
the relative symbolic link target was always computed
relative to the archive root. This leads to unresolvable
link target in case the symbolic link is not located in 
the archive root.
Instead, compute it relative to the symbolic link
parent directory.

Also, add a test for the case where the symbolic link
points into the archive but link target path is absolute and must be
converted to a relative path.

Bug: 429809
Change-Id: I88442305a62934a90bc8570beef96fbb63c79a88
Signed-off-by: Tobias Heide <tucks@gmx.de>
Also-by: Jan Sievers <jan.sievers@sap.com>