435068 Allow ignoring execution environments in dependency resolution
Add a new target platform configuration option
resolveWithExecutionEnvironmentConstraints which makes Tycho check that
the bundle being built and all its dependencies can resolve & run in an
OSGi container with the configured execution environment. This is the
current behaviour of Tycho, so the switch is true by default.

When the new switch is set to false, the p2 and Equinox resolver will
use all known execution environments: 

<plugin>
 <groupId>org.eclipse.tycho</groupId>
 <artifactId>target-platform-configuration</artifactId>
 <version>${tycho-version}</version>
 <configuration>
  <resolveWithExecutionEnvironmentConstraints>false
   </resolveWithExecutionEnvironmentConstraints>
 </configuration>
</plugin>

This allows required bundles to also use packages of other execution
environments (which otherwise leads to a p2 resolution error, cf. bug
434959) or declare a different BREE (which otherwise leads to missing
fragments, cf. bug 413116)

Added integration tests for both the p2 and Equinox resolver scenario.

Bug: 435068
Bug: 413116
Bug: 434959
Change-Id: I5dc3dd20d459a30c481928d059815fa215bc6a09
Signed-off-by: Jan Sievers <jan.sievers@sap.com>
Also-by: Tobias Oberlies <tobias.oberlies@sap.com>