441538 avoid printing stacktrace in case of p2 resolution error
Introduce typed exception ResolutionException (instead of
generic RuntimeException) for p2 resolution errors.
Catch as far upü in the stack a possible
(TychoMavenLifecycleParticipant) and wrap it into
a MavenExecutionException which will avoid printing the 
stacktrace and just show the message as ERROR (unless 
user specifies -e on the commandline)

Remove TargetDefinitionResolutionException and use
ResolutionExceptioni instead. TODO could also subclass
but I did not see a significant difference.

TODO should we really use a checked exception (causes
pervasive throws declarations everywhere in the code)
An alternative is to subclass (unchecked) RuntimeException
and catch it in TychoMavenLifecycleParticipant.
(less explicit but avoids big refactoring for throws clauses)

Bug: 441538
Change-Id: I3ef73b287e2cc5a17811f95a535495d5ffe507a8
Signed-off-by: Jan Sievers <jan.sievers@sap.com>