Update patch set 2
Patch Set 2: Code-Review-1

I like the change in the productive code, but the logic should not be tested through integration tests. So please add unit tests which test the logic/precedence of the switches.

In fact, I don't think that we need an integration test for this at all, although there is arguably some integration aspect in how unset parameters are propagated to the mojo. Still, I think that it is sufficient to test this once manually.

Patch-set: 2
Reviewer: Gerrit User 489 <489@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=-1