Update patch set 1
Patch Set 1: Code-Review-1

(5 comments)

Since you already implemented it, I'd keep the additional "force" option ;-)

IMHO, quite a few things need to be changed in the change though: Apart for the individual comments, I don't like adding an integration test. The only new integration aspect is that the new parameter is parsed and passed on correctly. Once we have dropped the error-prone conversion from and to Boolean, this IMHO no longer justifies an integration test. Everything else can be unit tested, so you should do this if you think it's worth it.

Patch-set: 1
Label: Code-Review=-1