Update patch set 1
Patch Set 1:

Integration test has really helped me to design this patch. And a first iteration of the patch was failing and integration-test caught it. The failure was in the retrieval of the property from command-line.
I believe there is value for integration test here since there is room for errors, or at least, there was.

Anyway, I'll apply your suggestion to this patch, and I'll try to add a unit test.

Patch-set: 1
Label: Verified=0