432957 Do not fail build if maven plugin used is part of build
If a maven-plugin and an eclipse-plugin that uses that maven plugin get
built in the same reactor, the build should not fail. 
If a plugin cannot be resolved (PluginResolutionException) because it
was not built yet, a check is done if that plugin is part of the
projects in the maven session. If so, continue without an exception, if
not, (re)throw the ResolutionException.

Change-Id: I483b4c550c5c9a3b99eea6c1d012164a4f47fb57
Signed-off-by: Martin Schreiber <m.schreiber@bachmann.info>