407500 Adding an option to let BREE define the JDK used for the tests.
-adding an parameter 'useJDK' with values 'SYSTEM' (=default) or 'BREE'.
For 'BREE' the JDK will be selected from the maven toolchains.xml based
on the MANIFEST BREE (MANIFEST BREE name must match the toolchains JDK
id). 
-Moving out the toolchain selection from the OsgiCompilerMojo to provide
functionality to select the toolchain the same way for compiler and
surefilre mojo (ToolchainProvider Component).
-Adding a unit test with easymock to test the different scenarios.
Adding the easymock to the dependency management section of the parent
pom.

Change-Id: If14f769a255ba2d25c82a954072e1393f10f38e2
Signed-off-by: Martin Schreiber <m.schreiber@bachmann.info>