442922 - Support usage of properties in build.properties
Introducing a serivce that interpolates Strings based on the maven
properties from the system properties, sessions user properties, maven
project properties and the sessions settings. The service should hide
the maven complexity regarding interpolation and is used within the
RootPropertiesParser to interpolate the values of the build properties.
Its also used by the PublishProductMojo, since there was already such an
interpolator in place.

Change-Id: I6ca3c6f4dc5874d338f0616a94d2e59c7c8f23ae
Signed-off-by: Martin Schreiber <m.schreiber@bachmann.info>